+++ codec-http2/src/main/java/io/netty/handler/codec/http2/DefaultHttp2HeadersEncoder.java
@@ -89,7 +89,7 @@ public class DefaultHttp2HeadersEncoder implements Http2HeadersEncoder, Http2Hea
      */
     private final class Http2HeaderTableEncoder extends DefaultHttp2HeaderTableListSize implements Http2HeaderTable {
         @Override
-        public void maxHeaderTableSize(int max) throws Http2Exception {
+        public void maxHeaderTableSize(long max) throws Http2Exception {
             if (max < 0) {
                 throw connectionError(PROTOCOL_ERROR, "Header Table Size must be non-negative but was %d", max);
             }
@@ -102,7 +102,7 @@ public class DefaultHttp2HeadersEncoder implements Http2HeadersEncoder, Http2Hea
         }
 
         @Override
-        public int maxHeaderTableSize() {
+        public long maxHeaderTableSize() {
             return encoder.getMaxHeaderTableSize();
         }
     }
